fix account and character bans

+++ sql/base/auth_database.sql
@@ -88,8 +88,8 @@ DROP TABLE IF EXISTS `account_banned`;
 /*!40101 SET character_set_client = utf8 */;
 CREATE TABLE `account_banned` (
   `id` int(10) unsigned NOT NULL DEFAULT '0' COMMENT 'Account id',
-  `bandate` int(10) unsigned NOT NULL DEFAULT '0',
-  `unbandate` int(10) unsigned NOT NULL DEFAULT '0',
+  `bandate` bigint(20) NOT NULL DEFAULT '0',
+  `unbandate` bigint(20) NOT NULL DEFAULT '0',
   `bannedby` varchar(50) NOT NULL,
   `banreason` varchar(255) NOT NULL,
   `active` tinyint(3) unsigned NOT NULL DEFAULT '1',
