+++ sql/base/characters_database.sql
@@ -435,8 +435,8 @@ DROP TABLE IF EXISTS `character_banned`;
 /*!40101 SET character_set_client = utf8 */;
 CREATE TABLE `character_banned` (
   `guid` int(10) unsigned NOT NULL DEFAULT '0' COMMENT 'Global Unique Identifier',
-  `bandate` int(10) unsigned NOT NULL DEFAULT '0',
-  `unbandate` int(10) unsigned NOT NULL DEFAULT '0',
+  `bandate` bigint(20) NOT NULL DEFAULT '0',
+  `unbandate` bigint(20) NOT NULL DEFAULT '0',
   `bannedby` varchar(50) NOT NULL,
   `banreason` varchar(255) NOT NULL,
   `active` tinyint(3) unsigned NOT NULL DEFAULT '1',
